Extract class CreateHttpImposterTests
